Move ToasterTestHelper.each_orderable_toppings ⚠️
This moves the class helper method `each_orderable_topping`, which is
used to generate test cases, into ToasterTestHelper.

----

This fails, because in this new context, `self` is ToasterTestHelper's
singleton class, instead of ToasterTest's singleton class. As such, the
method lookup fails, because ToasterTestHelper's singleton class is not
in ToasterTest's ancestor chain.