Extract ToasterTestHelper::ClassMethods
Since we want the `each_orderable_toppings` method to be available to
the `ToasterTest` class, we just need to make sure it's defined on an
object in it's ancestors.

Rather than define it on its singleton class, we instead define it on a
separate module, and extend the base with that module, adding it to the
singleton class' ancestors, accomplishing what we want. This is more
efficient, since we don't define the method each time the module is
included, and instead only define it once on the ClassMethods module.